Implemented proper ticking and fps counter for QT platform.

+++ src/core/platforms/StormPlatform.cpp
@@ -49,10 +49,10 @@ void StormPlatform::startMainLoop() {
             /* Just makes sure we dont devide by zero */
             deltaTime = 0.0001f;
         }
-        averageFps += 1000 / (1000 * deltaTime);
+        averageFps += 1000.0f / (1000.0f * deltaTime);
         averageFpsCounter++;
         if (averageFpsCounter >= 100) {
-            LOG(INFO) << "FPS: " << (float)averageFps / (float)averageFpsCounter;
+            LOG(INFO) << "FPS: " << averageFps / (float)averageFpsCounter;
             averageFpsCounter = 0;
             averageFps = 0.0f;
         }
